requirements.txt
numpy
mouse
pyautogui
mediapipe
pyalsaaudio
